+++ index.php
@@ -8,25 +8,22 @@ get_template_part('breadcrumbs'); ?>
 		$imageSize = $blog_layout == "fullwidth" ? 'awada_blog_full_thumb' : 'awada_blog_sidebar_thumb';
 		if ($blog_layout == "leftsidebar") {
 			get_sidebar();
-			$float = 8;
+			$float = 9;
 		} elseif ($blog_layout == "fullwidth") {
 			$float = 12;
 		} elseif ($blog_layout == "rightsidebar") {
-			$float = 8;
+			$float = 9;
 		} else {
 			$blog_layout = "rightsidebar";
-			$float = 8;
+			$float = 9;
 		} ?>
 		<div id="content" class="col-lg-<?php echo $float; ?> col-md-<?php echo $float; ?> col-sm-12 col-xs-12">
 			<div class="row">
 			   <div class="blog-masonry">
 					<div class="col-lg-12">
 						<?php
-						$paged = (get_query_var('paged')) ? get_query_var('paged') : 1;
-						$args = array('post_type' => 'post', 'paged' => $paged);
-						$wp_query = new WP_Query($args);
-						while ($wp_query->have_posts()):
-							$wp_query->the_post();
+						while (have_posts()):
+							the_post();
 							get_template_part('blog', 'content');
 						endwhile;
 						wp_link_pages(); ?>